make lampjs-ssr package, move ssr test to use package, update lampjs exports, switch to nodenext module resolution (might change back later or something)
