+++ lib-ssr/README.md
@@ -1,0 +1,23 @@
+# LampJs SSR
+
+LampJs SSR is a server-side rendering runtime for the LampJs framework. See [LampJs Documentation](https://github.com/JacksonO123/lampjs/blob/main/lib/docs.md)
+
+<br>
+
+## Create LampJs project
+
+To create a new LampJs project run
+
+pnpm:
+
+```sh
+pnpm create lampjs-app
+```
+
+npm:
+
+```sh
+npx create-lampjs-app
+```
+
+_By Jackson Otto_
