+++ lib-ssr/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "@jacksonotto/lampjs-ssr",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "description": "A server side rendering runtime for the LampJs framework",
+  "author": "Jackson Otto",
+  "version": "0.1.0",
+  "bin": {
+    "lampjs-ssr": "dist/server.js"
+  },
+  "scripts": {
+    "build": "npx tsc",
+    "watch": "npx nodemon -e ts --watch src --exec npx tsc",
+    "dev-start": "bun src/server.ts",
+    "start": "echo \"do something\""
+  },
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    },
+    "./server": {
+      "import": "./dist/server.js",
+      "types": "./dist/server.d.ts"
+    },
+    "./package.json": "./package.json"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.21",
+    "@types/node": "latest",
+    "nodemon": "latest",
+    "typescript": "latest"
+  },
+  "keywords": [
+    "lampjs",
+    "frontend",
+    "backend",
+    "javascript",
+    "typescript",
+    "framework",
+    "jsx",
+    "ssr"
+  ],
+  "dependencies": {
+    "express": "^4.18.2",
+    "vite": "^5.0.2"
+  },
+  "peerDependencies": {
+    "@jacksonotto/lampjs": "link:../lib"
+  }
+}
