+++ ssr/package.json
@@ -3,19 +3,17 @@
   "version": "0.1.0",
   "type": "module",
   "scripts": {
-    "start": "npx nodemon -e '*' --watch src --watch pkg --exec bun run pkg/server.ts",
+    "dev": "npx nodemon -e '*' --watch src --watch pkg --exec bun run server.ts",
     "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "devDependencies": {
     "@types/node": "latest",
-    "typescript": "latest",
-    "vite": "latest"
+    "typescript": "latest"
   },
   "dependencies": {
     "@jacksonotto/lampjs": "link:../lib",
-    "@types/express": "^4.17.21",
-    "express": "^4.18.2",
+    "@jacksonotto/lampjs-ssr": "link:../lib-ssr",
     "nodemon": "^3.0.1"
   }
 }
