+++ ssr/server.ts
@@ -1,0 +1,5 @@
+// import { startServer } from '/Users/jacksonotto/code/lampjs/lib-ssr/dist/server.js';
+import { startServer } from '@jacksonotto/lampjs-ssr/server';
+import App from './src/App.js';
+
+startServer(App);
