+++ ssr/src/App.tsx
@@ -1,13 +1,12 @@
+import { ServerSuspense } from '@jacksonotto/lampjs-ssr';
 // import { ServerSuspense } from '../pkg';
-import { ServerRouter } from '../pkg';
-// import Test from './Test';
-// import Async from './Async';
-import { Route } from '@jacksonotto/lampjs';
+import Test from './Test.js';
+import Async from './Async.js';
 
 const App = () => {
-  // const handleClick = () => {
-  //   console.log('clicked');
-  // };
+  const handleClick = () => {
+    console.log('clicked');
+  };
 
   return (
     <html>
@@ -19,24 +18,16 @@ const App = () => {
         <title>test</title>
       </head>
       <body>
-        <ServerRouter>
-          <Route path="/">
-            <h1>home</h1>
-            in body pt2.3
-            {/* <br /> */}
-            {/* <Test onClick={handleClick} /> */}
-            {/* <ServerSuspense */}
-            {/*   fallback={<span>waiting</span>} */}
-            {/*   waitServer */}
-            {/*   suspenseId="test" */}
-            {/* > */}
-            {/*   <Async /> */}
-            {/* </ServerSuspense> */}
-          </Route>
-          <Route path="/more">
-            <h1>more things</h1>
-          </Route>
-        </ServerRouter>
+        in body pt2.3
+        <br />
+        <Test onClick={handleClick} />
+        <ServerSuspense
+          fallback={<span>waiting</span>}
+          waitServer
+          suspenseId="test"
+        >
+          <Async />
+        </ServerSuspense>
       </body>
     </html>
   );
