+++ ssr/src/main.tsx
@@ -1,4 +1,5 @@
-import { mountSSR } from '../pkg';
-import App from './App';
+import { mountSSR } from '@jacksonotto/lampjs-ssr';
+// import { mountSSR } from '../pkg';
+import App from './App.js';
 
 mountSSR(document.body, <App />);
