+++ ssr/tsconfig.json
@@ -2,9 +2,9 @@
   "compilerOptions": {
     "target": "ESNext",
     "useDefineForClassFields": true,
-    "module": "ESNext",
+    "module": "NodeNext",
     "lib": ["ESNext", "DOM"],
-    "moduleResolution": "Node",
+    "moduleResolution": "NodeNext",
     "strict": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
@@ -19,5 +19,4 @@
     "jsxFragmentFactory": "Fragment",
     "allowSyntheticDefaultImports": true
   }
-  // "include": ["src"]
 }
